drm/exynos: gsc: normalize invalid buf idx from userspace
At least in the one frame processing mode which is the only mode supported
by the current gsc driver, the buf idx is not meaningful for the driver.
Because only one address in the buffer is valid at a time, so it is OK to
convert the invalid buf idx from userspace to a valid idx, process the
frame, and return the result with original buf idx.

Change-Id: Ibf07d2c31900ecb43bf17a39d616078cebb17c44
Signed-off-by: Hyungwon Hwang <human.hwang@samsung.com>